@@ -26,7 +26,7 @@ class ProjectAKVWrapper extends ComponentAKVWrapper
         }
     }
 
-    public function getPrefix(): string
+    public static function getPrefix(): string
     {
         return 'KBC::ProjectSecureKV::';
     }
