@@ -4,16 +4,15 @@ declare(strict_types=1);
 
 namespace Keboola\ObjectEncryptor\Tests;
 
+use Keboola\ObjectEncryptor\EncryptorOptions;
 use Keboola\ObjectEncryptor\Exception\ApplicationException;
 use Keboola\ObjectEncryptor\Exception\UserException;
-use Keboola\ObjectEncryptor\Wrapper\ConfigurationKMSWrapper;
 use Keboola\ObjectEncryptor\Wrapper\ProjectAKVWrapper;
 use Keboola\ObjectEncryptor\Wrapper\ProjectKMSWrapper;
 use Keboola\ObjectEncryptor\Wrapper\ProjectWideAKVWrapper;
 use Keboola\ObjectEncryptor\Wrapper\ProjectWideKMSWrapper;
-use PHPUnit\Framework\TestCase;
 
-class ProjectWideWrapperTest extends TestCase
+class ProjectWideWrapperTest extends AbstractTestCase
 {
     public function setUp(): void
     {
@@ -30,14 +29,17 @@ class ProjectWideWrapperTest extends TestCase
      */
     public function wrapperProvider(): array
     {
-        $projectWrapperKMS = self::createPartialMock(ProjectWideKMSWrapper::class, ['getRetries']);
-        $projectWrapperKMS->method('getRetries')->willReturn(1);
+        $projectWrapperKMS = new ProjectWideKMSWrapper(new EncryptorOptions(
+            stackId: 'some-stack',
+            kmsKeyId: self::getKmsKeyId(),
+            kmsRegion: self::getKmsRegion(),
+            backoffMaxTries: 1,
+        ));
 
-        $projectWrapperKMS->setKMSRegion((string) getenv('TEST_AWS_REGION'));
-        $projectWrapperKMS->setKMSKeyId((string) getenv('TEST_AWS_KMS_KEY_ID'));
-
-        $projectWrapperAKV = new ProjectWideAKVWrapper();
-        $projectWrapperAKV->setKeyVaultUrl((string) getenv('TEST_KEY_VAULT_URL'));
+        $projectWrapperAKV = new ProjectWideAKVWrapper(new EncryptorOptions(
+            stackId: 'some-stack',
+            akvUrl: self::getAkvUrl(),
+        ));
 
         return [
             'KMS' => [
@@ -56,7 +58,6 @@ class ProjectWideWrapperTest extends TestCase
     public function testEncrypt($wrapper): void
     {
         $secret = 'mySecretValue';
-        $wrapper->setStackId('my-stack');
         $wrapper->setProjectId('my-project');
         $encrypted = $wrapper->encrypt($secret);
         self::assertNotEquals($secret, $encrypted);
@@ -69,45 +70,37 @@ class ProjectWideWrapperTest extends TestCase
      */
     public function testEncryptDifferentProject($wrapper): void
     {
-        $wrapper->setStackId('my-stack');
         $wrapper->setProjectId('my-project');
         $secret = 'mySecretValue';
         $encrypted = $wrapper->encrypt($secret);
         self::assertNotEquals($secret, $encrypted);
         self::assertEquals($secret, $wrapper->decrypt($encrypted));
 
-        $wrapper->setStackId('my-stack');
         $wrapper->setProjectId('some-other-project');
         self::expectException(UserException::class);
         self::expectExceptionMessage('Deciphering failed.');
         $wrapper->decrypt($encrypted);
     }
 
-    public function testInvalidSetupEncryptKMS(): void
+    public function testInvalidSetupKMS(): void
     {
-        $wrapper = new ProjectKMSWrapper();
         self::expectException(ApplicationException::class);
         self::expectExceptionMessage('Cipher key settings are missing.');
-        $wrapper->encrypt('mySecretValue');
+        new ProjectKMSWrapper(new EncryptorOptions(
+            stackId: 'some-stack',
+            akvUrl: 'some-url'
+        ));
     }
 
-    public function testInvalidSetupEncryptAKV(): void
+    public function testInvalidSetupAKV(): void
     {
-        $wrapper = new ProjectAKVWrapper();
         self::expectException(ApplicationException::class);
         self::expectExceptionMessage('Cipher key settings are invalid.');
-        $wrapper->encrypt('mySecretValue');
-    }
-
-    /**
-     * @dataProvider wrapperProvider
-     * @param ProjectWideKMSWrapper|ProjectWideAKVWrapper $wrapper
-     */
-    public function testInvalidSetupEncryptStack($wrapper): void
-    {
-        self::expectException(ApplicationException::class);
-        self::expectExceptionMessage('No stack id provided.');
-        $wrapper->encrypt('mySecretValue');
+        new ProjectAKVWrapper(new EncryptorOptions(
+            stackId: 'some-stack',
+            kmsKeyId: 'some-key',
+            kmsRegion: 'some-region',
+        ));
     }
 
     /**
@@ -116,37 +109,9 @@ class ProjectWideWrapperTest extends TestCase
      */
     public function testInvalidSetupEncryptProjectId($wrapper): void
     {
-        $wrapper->setStackId('my-stack');
         self::expectException(ApplicationException::class);
         self::expectExceptionMessage('No project id provided.');
         $wrapper->encrypt('mySecretValue');
-    }
-
-    public function testInvalidSetupDecryptKMS(): void
-    {
-        $wrapper = new ProjectKMSWrapper();
-        self::expectException(ApplicationException::class);
-        self::expectExceptionMessage('Cipher key settings are missing.');
-        $wrapper->decrypt('mySecretValue');
-    }
-
-    public function testInvalidSetupDecryptAKV(): void
-    {
-        $wrapper = new ProjectAKVWrapper();
-        self::expectException(ApplicationException::class);
-        self::expectExceptionMessage('Cipher key settings are invalid.');
-        $wrapper->decrypt('mySecretValue');
-    }
-
-    /**
-     * @dataProvider wrapperProvider
-     * @param ProjectWideKMSWrapper|ProjectWideAKVWrapper $wrapper
-     */
-    public function testInvalidSetupDecryptStack($wrapper): void
-    {
-        self::expectException(ApplicationException::class);
-        self::expectExceptionMessage('No stack id provided.');
-        $wrapper->decrypt('mySecretValue');
     }
 
     /**
@@ -155,7 +120,6 @@ class ProjectWideWrapperTest extends TestCase
      */
     public function testInvalidSetupDecryptProjectId($wrapper): void
     {
-        $wrapper->setStackId('my-stack');
         self::expectException(ApplicationException::class);
         self::expectExceptionMessage('No project id provided.');
         $wrapper->decrypt('mySecretValue');
